% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigfish.R
\name{bigfish}
\alias{bigfish}
\title{bigfish}
\usage{
bigfish(
  table,
  alpha = 0.05,
  sim_lower = 1000,
  sim_upper = 5000,
  verbose = TRUE,
  visual = TRUE
)
}
\arguments{
\item{table}{an rXc contingency table (for 2X2 use twofish)}

\item{alpha}{A confidence coefficient (default = .05)}

\item{sim_lower}{A lower bound for number of iterations used to simulate p-value (default = 1000)}

\item{sim_upper}{An upper bound for number of iterations used to simulate p-value (default = 2000)}

\item{verbose}{A Boolean for whether or not to print extra information (default = TRUE)}

\item{visual}{A Boolean for whether or not to construct plots (default = TRUE)}
}
\value{
an invisible list called 'bf' that contains various tables,
a vector of simulated p-values,
the mean p-value from simulation
}
\description{
bigfish
}
\examples{
bigfish(table = table(c(1,2,3),c(1,2,3)))
}
