% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\docType{data}
\name{bond}
\alias{bond}
\title{bond}
\format{
## `bond`
A data frame with 10 rows and 3 columns:
\describe{
  \item{ARIScore}{integers ranging from 1-5 (a category)}
  \item{ADHESIVE}{Smartbond or Composite}
  \item{NUMBER}{integer values}
  ...
}
}
\source{
Mendenhall & Sinich Supplementary Data
}
\usage{
bond
}
\description{
Test Data for bigfish()
two way categorical data
10 observations of 3 variables
}
\keyword{datasets}
