% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapify.R
\name{bootstrapify}
\alias{bootstrapify}
\title{Bootstrapify}
\usage{
bootstrapify(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{vector from which the function samples}

\item{fun}{function used for point estimate}

\item{alpha}{value for confidence interval}

\item{cx}{scaling factor for graphics}

\item{...}{additional parameters}
}
\value{
an invisible list containing {x, a confidence interval, and the function} and a histogram
}
\description{
Bootstrapify
}
\examples{
bootstrapify(iter=10000,x = 1:100,fun="mean",alpha=0.05,cx=1.5)
}
