% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_diff_proportion_CIs.R
\name{cell_diff_proportion_CIs}
\alias{cell_diff_proportion_CIs}
\title{Helper function for computing the difference in cell proportions within my1way()}
\usage{
cell_diff_proportion_CIs(table, confidence = 0.95, adjusted = FALSE)
}
\arguments{
\item{table}{a one way table}

\item{confidence}{the desired level of confidence}

\item{adjusted}{whether or not to do Bonferoni adjustment}
}
\value{
nothing - only prints to console
}
\description{
Helper function for computing the difference in cell proportions within my1way()
}
\examples{
cell_diff_proportion_CIs(table = table(c(1,2),c(1,2)))
}
