% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data for fish}
\format{
## `ddt`
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{3 Letter river designation}
  \item{MILE}{milemarker of river}
  \item{SPECIES}{fish species}
  \item{LENGTH}{length of fish (cm)}
  \item{WEIGHT}{weight of fish (grams)}
  \item{DDT}{ddt levels in flesh of fish (ppm)}
  ...
}
}
\source{
<unknown>
}
\usage{
ddt
}
\description{
144 samples of fish
Categories: River, mile, species
Measurements of length, weight, and ddt levels
}
\keyword{datasets}
