% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\docType{data}
\name{hybrid}
\alias{hybrid}
\title{Test Data for mycontingency()}
\format{
## `hybrid`
A data frame with 4 rows and 3 columns:
\describe{
  \item{Model}{Conventional or Hybrid}
  \item{Claim}{Injury or No Injury}
  \item{Number}{integer values}
  ...
}
}
\source{
Mendenhall & Sinich Supplementary Data
}
\usage{
hybrid
}
\description{
two way categorical data
4 observations of 3 variables
}
\keyword{datasets}
