% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{logbin}
\usage{
logbin(x = 5, param = seq(0, 1, length = 1000), n = 20)
}
\arguments{
\item{x}{successes}

\item{param}{parameter of interest (p in this case)}

\item{n}{number of trials}
}
\value{
none
}
\description{
logbin
}
\examples{
logbin()
}
