% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1way.R
\name{my1way}
\alias{my1way}
\title{my1way}
\usage{
my1way(
  table,
  alpha = 0.05,
  adjusted = FALSE,
  test = TRUE,
  verbose = TRUE,
  visual = TRUE,
  screed = FALSE
)
}
\arguments{
\item{table}{a one way table}

\item{alpha}{significance value}

\item{adjusted}{Boolean for whether or not to Bonferonni correct}

\item{test}{Boolean for whether or not to perform Chi-squared test}

\item{verbose}{Boolean for whether or not to print extra information}

\item{visual}{Boolean for whether or not to show plots}

\item{screed}{Boolean for extra text dump of covariance matrices}
}
\value{
a table
}
\description{
my1way
}
\examples{
my1way(table = table(c(1,2,1,2)))
}
