% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of success}
}
\value{
a table and a bar plot
}
\description{
Binomial Simulation
}
\examples{
mybin()
}
