% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem (Poisson)}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of samples}

\item{lambda}{rate parameter for poisson distribution}

\item{...}{unnamed arguments}
}
\value{
1) a histogram of sample mean, 2) a bar plot of sampled y, and 3) a plot of the probability distribution for poisson
}
\description{
Central Limit Theorem (Poisson)
}
\examples{
mycltp(n=20, iter = 10000, lambda = 4)
}
