% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My maximum likelihood graphical solution}
\usage{
mymaxlik(
  lfun = logbin,
  x = c(1, 2, 3, 4, 5),
  param = seq(0, 1, length = 1000),
  ...
)
}
\arguments{
\item{lfun}{log likelihood function}

\item{x}{results of experiment}

\item{param}{parameter of interest}

\item{...}{this is a desription}
}
\value{
a list and a plot
}
\description{
My maximum likelihood graphical solution
}
\examples{
mymaxlik(lfun = logbin ,x = c(1,2,3,4,5), param = seq(0,1,length=1000))
}
