% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurve.R
\name{ncurve}
\alias{ncurve}
\title{My Normal Curve}
\usage{
ncurve(mu = 0, sigma = 1, a = 0)
}
\arguments{
\item{mu}{- mean of a normal distribution}

\item{sigma}{- standard deviation of a normal distribution}

\item{a}{- P(Y <= a)}
}
\value{
curve with area shaded and a probability
}
\description{
My Normal Curve
}
\examples{
myncurve()
}
