% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeOverbook.R
\name{optimizeOverbook}
\alias{optimizeOverbook}
\title{optimizeOverbook}
\usage{
optimizeOverbook(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{The total number of seats on a flight}

\item{gamma}{The pain tolerance}

\item{p}{The probability that a ticket holder shows up}
}
\value{
A named list, a graph of discrete case, a graph of continuous approximation
}
\description{
── R CMD check results ────────────────────────────────────────────── MATH4753RAMSSU23 0.1.0 ────
Duration: 15s
}
\details{
0 errors ✔ | 0 warnings ✔ | 0 notes ✔

R CMD check succeeded
}
\examples{
optimizeOverbook()
}
