% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseLinear.R
\name{piecewiseLinear}
\alias{piecewiseLinear}
\title{Piecewise Linear}
\usage{
piecewiseLinear(x, coef, k)
}
\arguments{
\item{x}{vector of data}

\item{coef}{vector of beta hats}

\item{k}{knot}
}
\value{
nothing
}
\description{
Piecewise Linear
}
\examples{
piecewiseLinear(1:10,c(1,2,3),4)
}
