% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\docType{data}
\name{scan}
\alias{scan}
\title{scan}
\format{
## `scan`
A data frame with 4 rows and 3 columns:
\describe{
  \item{RESPONSE}{observation index}
  \item{CATEGORY}{AgreeStrong or DisagreeStrong}
  \item{NUMBER}{integer values}
  ...
}
}
\source{
Mendenhall & Sinich Supplementary Data
}
\usage{
scan
}
\description{
Test Data for my1way
one way categorical data
4 observations of 3 variables
}
\keyword{datasets}
