% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHyper.R
\name{simHyper}
\alias{simHyper}
\title{Simulating Hypergeometric Distribution}
\usage{
simHyper(iter = 100, N = 20, r = 10, n = 5)
}
\arguments{
\item{iter}{number of iterations}

\item{N}{size of population}

\item{r}{number of successes in population}

\item{n}{sample size}
}
\value{
a table and a bar plot
}
\description{
Simulating Hypergeometric Distribution
}
\examples{
simHyper()
}
