% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofish.R
\name{twofish}
\alias{twofish}
\title{twofish - Fisher's Exact Test on 2X2 Contingency Tables}
\usage{
twofish(table, alpha = 0.05, alt = "greater", verbose = TRUE, visual = TRUE)
}
\arguments{
\item{table}{a 2X2 contingency table}

\item{alpha}{a confidence coefficient (default = .05)}

\item{alt}{the type of alternative hypothesis {'less', 'two.sided', 'greater'} ... default = two.sided}

\item{verbose}{A Boolean for whether or not to print extra information (default = TRUE)}

\item{visual}{A Boolean for whether or not to construct plots (default = TRUE)}
}
\value{
an invisible list called 'tf' that contains various tables,
the calculated p-value,and the sample estimate of the odds ration
}
\description{
twofish - Fisher's Exact Test on 2X2 Contingency Tables
}
\examples{
twofish(table = table(c(1,2),c(1,2)))
}
